mcp/pyproject.toml
[project]
name = "threat-detect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "nest-asyncio>=1.6.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.11",
    "typing-extensions>=4.15.0",
    "uvicorn[standard]>=0.35.0",
]
